pyproject.toml
[tool.poetry]
name = "identity"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-environ = "^0.11.2"
psycopg = "^3.2.3"
redis = "^5.1.0"
sentry-sdk = "^2.16.0"
dbt-copilot-python = "^0.2.2"
dj-database-url = "^2.2.0"
granian = "^1.6.3"
django-staff-sso-client = "^4.3.0"
django-ninja = "^1.3.0"
whitenoise = "^6.8.2"
django-simple-history = "^3.7.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pytest-django = "^4.9.0"
django-extensions = "^3.2.3"
coverage = "^7.6.4"
mypy = "^1.12.1"
pytest-mock = "^3.14.0"

[tool.poetry.group.docs.dependencies]
mkdocs-awesome-pages-plugin = "^2.8.0"
mkdocs-material = "^9.1.5"
mkdocs-minify-plugin = "^0.6.4"
mkdocs-redirects = "^1.2.0"
mkdocstrings-python = "^0.9.0"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "--disable-warnings --ds=config.settings.test"
DJANGO_SETTINGS_MODULE = "config.settings.test"

[tool.black]
target_version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
skip_gitignore = true
line_length = 88
lines_after_imports = 2

[tool.mypy]
exclude = [
    # TOML literal strings (single-quotes, no escaping necessary)
    '/migrations/',
]
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "django.*",
    "environ.*",
    "dbt_copilot_python.*",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = ["*/migrations/*", "*/tests/*", "config/*"]
